Improved KRC-20 envelope parsing
