+++ src/components/TransactionInfo.js
@@ -11,6 +11,7 @@ import {getTransaction, getTransactions} from '../kaspa-api-client.js';
 import BlueScoreContext from "./BlueScoreContext.js";
 import CopyButton from "./CopyButton.js";
 import PriceContext from "./PriceContext.js";
+import {parseSignatureScript} from "../inscriptions";
 
 const getOutputFromIndex = (outputs, index) => {
     for (const output of outputs) {
@@ -18,71 +19,6 @@ const getOutputFromIndex = (outputs, index) => {
             return output
         }
     }
-}
-
-const hexToBytes = (hex) => {
-    const bytes = [];
-    for (let i = 0; i < hex.length; i += 2) {
-        bytes.push(parseInt(hex.substring(i, i + 2), 16));
-    }
-    return new Uint8Array(bytes);
-};
-
-const bytesToHex = (bytes) => {
-    return Array.from(bytes).map(b => b.toString(16).padStart(2, '0')).join('');
-}
-
-const bytesToString = (bytes) => {
-    return new TextDecoder().decode(bytes);
-}
-
-const parseSignatureScript = (hex, inner) => {
-    /* Attempt at decoding Kasplex deploy&mint */
-    /* Example deploy: 105424172e946f85daf87f50422e4a5a7f73d541a7a0eaf666cf40567a80ba5d */
-    /* Example mint: e2792d606f7cae9994cbc233a2cdb9c4d4541ad195852ae8ecfd787254613ded */
-    const bytes = hexToBytes(hex);
-    let offset = 0;
-    const result = [];
-    let keyFound = false;
-
-    while (offset < bytes.length) {
-        const opcode = bytes[offset];
-        offset += 1;
-        if (opcode >= 0x01 && opcode <= 0x4b) {
-            const dataLength = opcode;
-            const data = bytes.slice(offset, offset + dataLength);
-            offset += dataLength;
-            if (!keyFound) {
-                // result.push(`PUB_KEY ${bytesToHex(data)}`);
-                keyFound = true;
-            } else {
-                result.push(`OP_PUSH ${bytesToString(data)}`);
-            }
-        } else if (opcode === 0x4c) {
-            const dataLength = bytes[offset];
-            offset += 1;
-            const data = bytes.slice(offset, offset + dataLength);
-            if (data[0] !== 0x7b) { // FIXME: Need a better selector on when to recurse
-                result.push(parseSignatureScript(bytesToHex(data), true));
-            } else {
-                result.push(`OP_PUSH ${bytesToString(data)}`)
-            }
-            offset += dataLength;
-        } else if (opcode === 0x00) {
-            // result.push('OP_0')
-        } else if (opcode === 0x51) {
-            // result.push('OP_1')
-        } else if (opcode === 0x63) {
-            // result.push('OP_IF')
-        } else if (opcode === 0x68) {
-            // result.push('OP_ENDIF')
-        } else if (opcode === 0xac) {
-            // result.push('OP_CHECKSIG')
-        } else {
-            // result.push(`UNKNOWN ${opcode.toString(16)}`)
-        }
-    }
-    return result.join('\n');
 }
 
 const TransactionInfo = () => {
@@ -273,9 +209,9 @@ const TransactionInfo = () => {
                                                 </div>
                                                 {tx_input.signature_script.includes('6b6173706c6578') && (
                                                     <>
-                                                        <div className="blockinfo-key mt-2">Signature Script Decode</div>
+                                                        <div className="blockinfo-key mt-2">KRC-20</div>
                                                         <div className="utxo-value-mono" style={{ whiteSpace: "pre-wrap" }}>
-                                                            {parseSignatureScript(tx_input.signature_script)}
+                                                            {parseSignatureScript(tx_input.signature_script).findLast(s => s.indexOf('OP_PUSH') === 0).split(' ')[1]}
                                                         </div>
                                                     </>
                                                 )}
